Create robots.txt file
The “User-agent: *” part means that it applies to all robots. The “Disallow: /” part means that it applies to your entire website.
In effect, this will tell all robots and web crawlers that they are not allowed to access or crawl your site.

Disallowing all robots on a live website can lead to your site being removed from search engines and can result in a loss of traffic and revenue.